- struct "Item": add variables <index, UUID>
- class "List Manager": adjust each item's variables <index>

- add class "Data Manager" for data persistence:
	+ still get error where data can't be decoded

+++ Listing/Model/ListManager.swift
@@ -10,9 +10,15 @@ import Foundation
 
 class ListManager {
     
-    private var itemsInList: [Item] = [
-        Item(title: "Tap to Delete")
+    var itemsInList: [Item] = [
+        Item(title: "Tap to Delete", index: 0, itemIdentifier: UUID())
     ]
+    
+    var newList: [Item] {
+        return itemsInList.sorted { (item1, item2) -> Bool in
+            item1.index < item2.index
+        }
+    }
     
     var listCount: Int {
         return itemsInList.count
@@ -21,12 +27,17 @@ class ListManager {
     func itemAtIndex(_ index: Int) -> Item {
         return itemsInList[index]
     }
-    
-    func addItem(_ item: Item) {
-        itemsInList.append(item)
-    }
-    
+
     func addItemAtFirst(_ item: Item) {
+        if listCount == 1 {
+            saveNewItem(ofIndex: 0, by: +1)
+        } else if listCount > 1 {
+            for i in 0...listCount-1 {
+//                saveNewItem(ofIndex: i, by: +1)
+                itemsInList[i].changeIndex(by: 1)
+            }
+        }
+        item.save()
         itemsInList.insert(item, at: 0)
     }
     
@@ -35,12 +46,48 @@ class ListManager {
     }
     
     func deleteItem(at index: Int) {
+        if listCount > 1 {
+            for i in 0...listCount - 1 {
+                if itemsInList[i].index > index {
+//                    saveNewItem(ofIndex: i, by: -1)
+                    itemsInList[i].changeIndex(by: -1)
+                }
+            }
+        }
+        itemsInList[index].delete()
         itemsInList.remove(at: index)
     }
     
-    func moveItem(_ item: Item, from startIndex: Int, to endIndex: Int) {
+    func moveItem(from startIndex: Int, to endIndex: Int) {
+        guard startIndex != endIndex else { return }
+        
+        var item = itemAtIndex(startIndex)
+//        var holerItem = item
+//        item.delete()
+//        holerItem.index = endIndex
+//        let newItem = holerItem
+//        newItem.save()
+        item.index = endIndex
+        
         deleteItem(at: startIndex)
+        
+        for i in 0...listCount - 1 {
+            if itemsInList[i].index >= endIndex {
+//                saveNewItem(ofIndex: i, by: +1)
+                itemsInList[i].changeIndex(by: 1)
+            }
+        }
+        
         itemsInList.insert(item, at: endIndex)
+    }
+    
+    func saveNewItem(ofIndex i: Int, by value: Int) {
+        let oldItem = itemsInList[i]
+        var newItem = oldItem
+        newItem.changeIndex(by: value)
+        newItem.save()
+        replaceItem(with: newItem, at: i)
+        oldItem.delete()
     }
 
 }
